Move rename and value to the vdom file
