+++ packages/block-library/src/navigation/vdom.js
@@ -2,10 +2,6 @@
  * External dependencies
  */
 import { h } from 'preact';
-/**
- * Internal dependencies
- */
-import { rename, value } from './directives';
 
 // Recursive function that transfoms a DOM tree into vDOM.
 export default function toVdom( node ) {
@@ -42,3 +38,23 @@ export default function toVdom( node ) {
 
 // Filter existing items.
 const exists = ( x ) => x;
+
+// Rename WordPress Directives from:
+// - `wp-some-directive` to `someDirective`.
+// - `wp-some-prefix:some-directive` to `somePrefix`.
+export const rename = ( name ) =>
+	name
+		.toLowerCase()
+		.replace( /^wp-/, '' )
+		.replace( /:.*/, '' )
+		.replace( /-(.)/g, ( _, chr ) => chr.toUpperCase() );
+
+// Add some support for suffixes and modifiers. See the tests file.
+export const value = ( name, val ) => {
+	if ( ! name.includes( ':' ) ) return val;
+	return {
+		name: rename( name ),
+		suffix: rename( name.substring( name.indexOf( ':' ) + 1 ) ),
+		value: val,
+	};
+};
